requirements.txt
flwr
torch
opacus
scikit-learn
pandas
numpy
pyyaml
streamlit
matplotlib
plotly
pytest
